refactor(library/Parser): extracts "library/Parse"
