+++ src/library/Parse/exports.ts
@@ -1,0 +1,3 @@
+export {
+  Parse_instanceFrom as instanceFrom,
+} from './instanceFrom';
