+++ src/library/Parse/index.ts
@@ -1,0 +1,5 @@
+import * as Parse from './exports.ts';
+
+export {
+  Parse as default,
+};
