+++ src/library/Parser/exports.ts
@@ -2,4 +2,4 @@ export type * from './StaticParser';
 
 export type * from './Parsable';
 
-export * from './Parse_instanceFrom';
+export * from '../Parse/instanceFrom';
